Removed async from Commands and renamed namespace too, removed InputState system as it didn't even really work, changed how IDataContext stores data (string keys instead of DataKey keys), and added context capturing MenuItems for menus at the top of the window

+++ FramePFX/Utils/NumberAverager.cs
@@ -24,7 +24,7 @@ namespace FramePFX.Utils {
                 average += elem;
             }
 
-            return average / (double) this.averages.Length;
+            return average / this.averages.Length;
         }
     }
 }
